query: add ability to check for test results
Add an option to query.py to check for test results. In a later patch we
will send email notifications for test results. Keep a list of message
IDs for which notifications have already been processed in a new
database. Run this query as a new step in the workflow.

Signed-off-by: Vincent Fu <vincent.fu@samsung.com>