github actions: run workflow every 15 minutes
Set this loose in the wild. Automatically run the workflow every 15
minutes to query the mailing list for new patch series and query GitHub
for completed tests. When new patch series are found, push them to a new
branch. When completed tests are found, send a notification email to the
mailing list.

Signed-off-by: Vincent Fu <vincent.fu@samsung.com>